[Fix] 카테고리별 조회 수정
